backend/requirements.txt
Flask
pymongo
fastapi
uvicorn
motor 
pydantic 
uvicorn
crewai
crewai-tools
pdfplumber
paddleocr
reportlab
python-dotenv
